requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
config==0.5.1
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.13.4
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
textblob==0.18.0.post0
tokenizers==0.20.3
toml==0.10.2
torch==2.2.2
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
